% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks_table}
\alias{ks_table}
\title{Get KS statistics}
\usage{
ks_table(yhat, y)
}
\arguments{
\item{yhat}{numeric the value predicted by your model.}

\item{y}{numeric the target values.}
}
\value{
The data frame containing tpr, fpr and ks.
}
\description{
Output a \code{data.frame} with inputs (y, yhat) and tpr, fpr and ks.
}
\examples{
data("two_class_example_edited")
max(ks_table(two_class_example_edited$yhat, two_class_example_edited$y)$ks)
}
