% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_example_edited.R
\docType{data}
\name{two_class_example_edited}
\alias{two_class_example_edited}
\title{A sample with the target and predictions}
\format{A data frame with 500 rows and 2 variables:
\describe{
\item{yhat}{predictions}
\item{y}{target}
}}
\source{
package \code{yardstick::two_class_example}
}
\usage{
two_class_example_edited
}
\description{
A dataset containing two columns, \code{yhat} and \code{y} as
the inputs for the function \code{ks_table}.
}
\keyword{datasets}
